requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dateutil==2.9.0.post0
